[♻️Refactor]: 파일명 변경 및 함수&컴포넌트로 코드 분리
파일명을 변경했으며,
재사용성을 위해 중복되는 코드들은 함수 및 컴포넌트로 분리
